Bug#25111169 - ERROR DURING SERVER INITIALIZATION WITH INNODB_UNDO_TABLESPACES
undo::done_logging() is now used for undo tablespaces that are created during
bootstrap by the function srv_undo_tablespaces_construction_list_clear().
In this usage, the truncation log file does not exist.  The function
os_file_delete() assumes the file exists.

Add code to srv_undo_tablespaces_construction_list_clear() so that it only
calls undo::done_logging() if undo::is_truncate_log_present() returns true.
Also, fix undo::done_logging so that it does not attempt to delete a file
that does not exist.

Approved by Satya in RB#14629

+++ storage/innobase/srv/srv0start.cc
@@ -587,7 +587,7 @@ srv_undo_tablespace_fixup(
 {
 	undo::Tablespace	undo_space(space_id);
 
-	if (undo::is_truncate_log_present(space_id)) {
+	if (undo::is_active_truncate_log_present(space_id)) {
 
 		ib::info() << "Undo Tablespace number " << space_id
 			<< " was being truncated when mysqld quit.";
@@ -951,8 +951,10 @@ srv_undo_tablespaces_construction_list_clear()
 			space_id, BUF_REMOVE_FLUSH_WRITE, NULL);
 
 		/* Remove the truncate redo log file if it exists. */
-		undo::Truncate	undo_trunc;
-		undo_trunc.done_logging(space_id);
+		if (undo::is_active_truncate_log_present(space_id)) {
+			undo::Truncate	undo_trunc;
+			undo_trunc.done_logging(space_id);
+		}
 	}
 
 	undo::clear_construction_list();
